requirements.txt
pandas
numpy
pvlib
#tensorflow
pymongo
python-dotenv
pydantic
torch
scikit-learn
joblib